Update some names.
