+++ Engine/Controls/FineTuningJobListControl.xaml.cs
@@ -18,7 +18,7 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 	/// <summary>
 	/// Interaction logic for ProjectsListControl.xaml
 	/// </summary>
-	public partial class FineTuningJobListControl : UserControl, IBindData<FineTune>
+	public partial class FineTuningJobListControl : UserControl, IBindData<FineTuningItem>
 	{
 		public FineTuningJobListControl()
 		{
@@ -165,7 +165,7 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 
 		[Category("Main")]
 
-		public FineTune Data
+		public FineTuningItem Data
 		{
 			get => _Data;
 			set
@@ -184,7 +184,7 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 				MustRefresh = true;
 			}
 		}
-		public FineTune _Data;
+		public FineTuningItem _Data;
 
 		public bool MustRefresh;
