+++ Engine/Controls/FineTuningRemoteDataControl.xaml.cs
@@ -18,7 +18,7 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 	/// <summary>
 	/// Interaction logic for ProjectsListControl.xaml
 	/// </summary>
-	public partial class FineTuningRemoteDataControl : UserControl, IBindData<FineTune>
+	public partial class FineTuningRemoteDataControl : UserControl, IBindData<FineTuningItem>
 	{
 		public FineTuningRemoteDataControl()
 		{
@@ -168,7 +168,7 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 
 		[Category("Main")]
 
-		public FineTune Data
+		public FineTuningItem Data
 		{
 			get => _Data;
 			set
@@ -187,7 +187,7 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 				MustRefresh = true;
 			}
 		}
-		public FineTune _Data;
+		public FineTuningItem _Data;
 
 		public bool MustRefresh;
